+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This release has an [MSRV] of 1.82.
 - `Image` now stores the alpha as an `f32` ([#65][] by [@waywardmonkeys][])
 - Use `color` crate. See below for details ([#63][] by [@waywardmonkeys][])
 - `Gradient`, `Image`, `Brush` now have `with_alpha` and `Gradient` also gets a `multiply_alpha` ([#67][] by [@waywardmonkeys][])
+- `Gradient` now tracks a hue direction and interpolation color space ([#71][] by [@waywardmonkeys][])
 
 ### Color Changes
 
@@ -78,6 +79,7 @@ This release has an [MSRV] of 1.70.
 [#63]: https://github.com/linebender/peniko/pull/63
 [#65]: https://github.com/linebender/peniko/pull/65
 [#67]: https://github.com/linebender/peniko/pull/67
+[#71]: https://github.com/linebender/peniko/pull/71
 
 [@DJMcNab]: https://github.com/DJMcNab
 [@ratmice]: https://github.com/ratmice